Tobacco lag times

Prepare the disease specific functions that describe how a change in tobacco consumption gradually has an effect on the relative risk of disease incidence over time (up to 40 years) since e.g. someone quit smoking

TobLags(
  disease_name = c("Pharynx", "Oral_cavity"),
  n_years = 40,
  lag_data = tobalcepi::tobacco_lag_times
)

Arguments

disease_name

Character - the name of the disease under consideration.

n_years

Integer - the number of years from 1 to n over which the effect of a change in consumption emerges. Defaults to 40 years to fit with the current lag data.

lag_data

Data table containing the numerical description of the lag function. The data table "tobacco lag times" is embedded within the stapmr package.

Value

Returns a data table with two columns - one for the years since consumption changed, and the other that gives the proportion by which the effect of a change in consumption on an individual's relative risk of disease has so far emerged.

Details

All lag times are taken from a re-analysis of the Cancer prevention II study by Oza et al 2011 and Kontis et al 2014 The values were sent to us by Kontis. Lags are smoothed functions over time describing the proportion of the excess risk due to smoking that still remains.

Kontis et al. re-analysed the change in risk after smoking in the ACS-CPS II study from Oza et al., producing three functions to describe the decline in risk after quitting for each of cancers, CVD and COPD. The estimates were informed by data on former smokers with known quit dates who were disease-free at baseline. The results show the proportion of excess relative risk remaining at each time-point since cessation. A cross-check showed that the figures for cancers were broadly consistent with the findings of the International Agency for Research on Cancer's (IARC) 2007 review of the decline in risk after quitting smoking.

The remaining question is how risk declines after quitting smoking for diseases that are not cancers, CVD or COPD. Kontis et al. state that "Randomised trials also indicate that the benefits of behaviour change and pharmacological treatment on diabetes risk occur within a few years, more similar to the CVDs than cancers. Therefore, we used the CVD curve for diabetes." In-line with Kontis, we apply the rate of decline in risk of CVD after quitting smoking to type 2 diabetes. For other diseases, we assume that the relative risk reverts to 1 immediately after quitting i.e. an immediate rather than a gradual decline in risk.

Examples

TobLags("Pharynx")
#>     time_since_quit prop_risk_reduction
#>  1:               0          0.00000000
#>  2:               1          0.08876399
#>  3:               2          0.16923763
#>  4:               3          0.24226931
#>  5:               4          0.30861345
#>  6:               5          0.36894171
#>  7:               6          0.42385267
#>  8:               7          0.47388041
#>  9:               8          0.51950194
#> 10:               9          0.56114374
#> 11:              10          0.59918751
#> 12:              11          0.63397523
#> 13:              12          0.66581360
#> 14:              13          0.69497793
#> 15:              14          0.72171559
#> 16:              15          0.74624910
#> 17:              16          0.76877877
#> 18:              17          0.78948509
#> 19:              18          0.80853085
#> 20:              19          0.82606301
#> 21:              20          0.84221434
#> 22:              21          0.85710491
#> 23:              22          0.87084340
#> 24:              23          0.88352822
#> 25:              24          0.89524863
#> 26:              25          0.90608559
#> 27:              26          0.91611261
#> 28:              27          0.92539650
#> 29:              28          0.93399802
#> 30:              29          0.94197248
#> 31:              30          0.94937023
#> 32:              31          0.95623718
#> 33:              32          0.96261520
#> 34:              33          0.96854251
#> 35:              34          0.97405401
#> 36:              35          0.97918161
#> 37:              36          0.98395447
#> 38:              37          0.98839930
#> 39:              38          0.99254055
#> 40:              39          0.99640061
#> 41:              40          1.00000000
#>     time_since_quit prop_risk_reduction